Success Test case: Write routes for sweets crud operation and search operation
Create the route that is used to handle the sweet operations and also use Router
of fastapi to modularize the code in modules.

+++ server/src/models/sweets.py
@@ -18,7 +18,7 @@ class SweetModel(Document):
 
     name: str = Field(..., max_length=50)
     category: Link[CategoryModel]
-    price: int = Field(..., ge=0)
+    price: float = Field(..., ge=0)
     quantity: int = Field(..., ge=0)
 
     class Settings:
